fix(loader): size and margin
